+++ lib/Api/PaymentTransactionsApi.php
@@ -477,6 +477,107 @@ class PaymentTransactionsApi
     }
 
     /**
+     * Operation getPaymentTransactionsOldFormat
+     *
+     * GET Payment/Transactions/{paymentTransactionId}/OldFormat
+     *
+     * @param string $payment_transaction_id Payment ID (PCI_...) or hash (required)
+     * @throws ApiException on non-2xx response
+     * @return \Secuconnect\Client\Model\SecupayTransactionProductModel
+     */
+    public function getPaymentTransactionsOldFormat($payment_transaction_id)
+    {
+        list($response) = $this->getPaymentTransactionsOldFormatWithHttpInfo($payment_transaction_id);
+        return $response;
+    }
+
+    /**
+     * Operation getPaymentTransactionsOldFormatWithHttpInfo
+     *
+     * GET Payment/Transactions/{paymentTransactionId}/OldFormat
+     *
+     * @param string $payment_transaction_id Payment ID (PCI_...) or hash (required)
+     * @throws ApiException on non-2xx response
+     * @return array of \Secuconnect\Client\Model\SecupayTransactionProductModel, HTTP status code, HTTP response headers (array of strings)
+     */
+    public function getPaymentTransactionsOldFormatWithHttpInfo($payment_transaction_id)
+    {
+        // verify the required parameter 'payment_transaction_id' is set
+        if ($payment_transaction_id === null || (is_array($payment_transaction_id) && count($payment_transaction_id) === 0)) {
+            throw new \InvalidArgumentException(
+                'Missing the required parameter $payment_transaction_id when calling getPaymentTransactionsOldFormat'
+            );
+        }
+        // parse inputs
+        $resourcePath = "/Payment/Transactions/{paymentTransactionId}/OldFormat";
+        $httpBody = '';
+        $queryParams = [];
+        $headerParams = [];
+        $formParams = [];
+        $_header_accept = $this->apiClient->selectHeaderAccept(['application/json']);
+        if (!is_null($_header_accept)) {
+            $headerParams['Accept'] = $_header_accept;
+        }
+        $headerParams['Content-Type'] = $this->apiClient->selectHeaderContentType([]);
+
+        // path params
+        if ($payment_transaction_id !== null) {
+            $resourcePath = str_replace(
+                "{" . "paymentTransactionId" . "}",
+                $this->apiClient->getSerializer()->toPathValue($payment_transaction_id),
+                $resourcePath
+            );
+        }
+
+        // for model (json/xml)
+        if (isset($_tempBody)) {
+            $httpBody = $_tempBody; // $_tempBody is the method argument, if present
+        } elseif (count($formParams) > 0) {
+            $httpBody = $formParams; // for HTTP post (form)
+        }
+        for ($retries = 0; ; $retries++) {
+
+            // this endpoint requires OAuth (access token)
+            if (strlen($this->apiClient->getConfig()->getAccessToken()) !== 0) {
+                $headerParams['Authorization'] = 'Bearer ' . $this->apiClient->getConfig()->getAccessToken();
+            }
+
+            // make the API Call
+            try {
+                list($response, $statusCode, $httpHeader) = $this->apiClient->callApi(
+                    $resourcePath,
+                    'GET',
+                    $queryParams,
+                    $httpBody,
+                    $headerParams,
+                    '\Secuconnect\Client\Model\SecupayTransactionProductModel',
+                    '/Payment/Transactions/{paymentTransactionId}/OldFormat'
+                );
+
+                return [$this->apiClient->getSerializer()->deserialize($response, '\Secuconnect\Client\Model\SecupayTransactionProductModel', $httpHeader), $statusCode, $httpHeader];
+            } catch (ApiException $e) {
+                switch ($e->getCode()) {
+                    case 200:
+                        $data = $this->apiClient->getSerializer()->deserialize($e->getResponseBody(), '\Secuconnect\Client\Model\SecupayTransactionProductModel', $e->getResponseHeaders());
+                        $e->setResponseObject($data);
+                        break;
+                    case 401:
+                        if ($retries < 1) {
+                            Authenticator::reauthenticate();
+                            continue 2;
+                        }
+                    default:
+                        $data = $this->apiClient->getSerializer()->deserialize($e->getResponseBody(), '\Secuconnect\Client\Model\ProductExceptionPayload', $e->getResponseHeaders());
+                        $e->setResponseObject($data);
+                        break;
+                }
+
+                throw $e;
+            }
+        }
+    }
+
+    /**
      * Operation getShippingUrl
      *
      * GET Payment/Transactions/{paymentTransactionId}/shippingUrl